+++ apps/party/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gtmgrid/party
 
+## 1.3.0
+
+### Patch Changes
+
+- @gtmgrid/auth@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes