+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gtmgrid/auth
 
+## 1.3.0
+
+### Patch Changes
+
+- @gtmgrid/analytics@1.3.0
+- @gtmgrid/db@1.3.0
+- @gtmgrid/email@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes